Commit before attempting to combine ConnectedThread and BtSerial
